Only run deviating outputDir check if we don't find anything in the sourceDir
