+++ bin/tasks/build.js
@@ -5,7 +5,6 @@ import semver from 'semver';
 import tmp from 'tmp-promise';
 
 import { createTask, transitionTo } from '../lib/tasks';
-import deviatingOutputDir from '../ui/messages/warnings/deviatingOutputDir';
 import { initial, pending, skipped, success } from '../ui/tasks/build';
 
 export const setSourceDir = async ctx => {
@@ -40,16 +39,6 @@ export const setSpawnParams = ctx => {
   };
 };
 
-const getOutputDir = buildLog => {
-  const outputString = 'Output directory: ';
-  const outputIndex = buildLog.lastIndexOf(outputString);
-  if (outputIndex === -1) return undefined;
-  const remainingLog = buildLog.substr(outputIndex + outputString.length);
-  const newlineIndex = remainingLog.indexOf('\n');
-  const outputDir = newlineIndex === -1 ? remainingLog : remainingLog.substr(0, newlineIndex);
-  return outputDir.trim();
-};
-
 export const buildStorybook = async ctx => {
   ctx.buildLogFile = path.resolve('./build-storybook.log');
   const logFile = fs.createWriteStream(ctx.buildLogFile);
@@ -63,17 +52,6 @@ export const buildStorybook = async ctx => {
     await execa(command, [...clientArgs, ...scriptArgs], { stdio: [null, logFile, logFile] });
   } finally {
     logFile.end();
-  }
-
-  try {
-    const buildLog = fs.readFileSync(ctx.buildLogFile, 'utf8');
-    const outputDir = getOutputDir(buildLog);
-    if (outputDir && outputDir !== ctx.sourceDir) {
-      ctx.log.warn(deviatingOutputDir(ctx, outputDir));
-      ctx.sourceDir = outputDir;
-    }
-  } catch (e) {
-    ctx.log.debug(e);
   }
 };
