+++ bin/tasks/upload.js
@@ -1,11 +1,12 @@
 /* eslint-disable no-param-reassign */
-import { readdirSync, statSync } from 'fs';
+import { readdirSync, readFileSync, statSync } from 'fs';
 import { join } from 'path';
 import slash from 'slash';
 import { URL } from 'url';
 
 import { createTask, transitionTo } from '../lib/tasks';
 import uploadFiles from '../lib/uploadFiles';
+import deviatingOutputDir from '../ui/messages/warnings/deviatingOutputDir';
 import {
   failed,
   initial,
@@ -50,26 +51,57 @@ function getPathsInDir(rootDir, dirname = '.') {
   }
 }
 
+function getOutputDir(buildLog) {
+  const outputString = 'Output directory: ';
+  const outputIndex = buildLog.lastIndexOf(outputString);
+  if (outputIndex === -1) return undefined;
+  const remainingLog = buildLog.substr(outputIndex + outputString.length);
+  const newlineIndex = remainingLog.indexOf('\n');
+  const outputDir = newlineIndex === -1 ? remainingLog : remainingLog.substr(0, newlineIndex);
+  return outputDir.trim();
+}
+
+function getFileInfo(sourceDir) {
+  const lengths = getPathsInDir(sourceDir).map(o => ({ ...o, knownAs: slash(o.pathname) }));
+  const paths = lengths.map(({ knownAs }) => knownAs);
+  const total = lengths.map(({ contentLength }) => contentLength).reduce((a, b) => a + b, 0);
+  return { lengths, paths, total };
+}
+
+const isValidStorybook = ({ paths, total }) =>
+  total > 0 && paths.includes('iframe.html') && paths.includes('index.html');
+
 export const uploadStorybook = async (ctx, task) => {
-  const { client, sourceDir } = ctx;
+  let fileInfo = getFileInfo(ctx.sourceDir);
 
-  const pathAndLengths = getPathsInDir(sourceDir).map(o => ({ ...o, knownAs: slash(o.pathname) }));
-  const paths = pathAndLengths.map(({ knownAs }) => knownAs);
-  const total = pathAndLengths.map(({ contentLength }) => contentLength).reduce((a, b) => a + b, 0);
+  if (!isValidStorybook(fileInfo) && ctx.buildLogFile) {
+    try {
+      const buildLog = readFileSync(ctx.buildLogFile, 'utf8');
+      const outputDir = getOutputDir(buildLog);
+      if (outputDir && outputDir !== ctx.sourceDir) {
+        ctx.log.warn(deviatingOutputDir(ctx, outputDir));
+        ctx.sourceDir = outputDir;
+        fileInfo = getFileInfo(ctx.sourceDir);
+      }
+    } catch (e) {
+      ctx.log.debug(e);
+    }
+  }
 
-  if (!total || !paths.includes('iframe.html')) {
+  if (!isValidStorybook(fileInfo)) {
     throw new Error(invalid(ctx).output);
   }
 
   task.output = preparing(ctx).output;
 
-  const { getUploadUrls } = await client.runQuery(TesterGetUploadUrlsMutation, { paths });
+  const { lengths, paths, total } = fileInfo;
+  const { getUploadUrls } = await ctx.client.runQuery(TesterGetUploadUrlsMutation, { paths });
   const { domain, urls } = getUploadUrls;
   const files = urls.map(({ path, url, contentType }) => ({
-    path: join(sourceDir, path),
+    path: join(ctx.sourceDir, path),
     url,
     contentType,
-    contentLength: pathAndLengths.find(({ knownAs }) => knownAs === path).contentLength,
+    contentLength: lengths.find(({ knownAs }) => knownAs === path).contentLength,
   }));
 
   task.output = starting(ctx).output;
